`subliminal-test` now requires that the user specify a simulator version.
If a version is not specified, it is not clear which version of the simulator to reset.
We also can't guarantee the version of the simulator that will be used unless we set it,
given the variances in the behavior of `instruments` as described in
http://openradar.appspot.com/radar?id=3107401.

It also makes the README a bit simpler to only describe behavior that we can guarantee.